Make all top-level `pyrtl` names canonical, and stop using submodule names.
This makes the documentation consistently use top-level names like
`pyrtl.WireVector`, instead of submodule-level names like
`pyrtl.wire.WireVector`. This also updates `examples` and `tests` to
consistently use the canonical top-level names.

Users should not be aware of pyrtl's submodule structure. For example, users
should not know or care that `mux` is a `corecircuit` and not a `helperfunc`.